+++ xclim/core/units.py
@@ -201,7 +201,7 @@ def pint2cfunits(value: units.Quantity | units.Unit) -> str:
     if isinstance(value, (pint.Quantity, units.Quantity)):
         value = value.units
 
-    return f"{value:~cf}"
+    return f"{value:cf}"
 
 
 def ensure_cf_units(ustr: str) -> str: